@@ -9,8 +9,14 @@
   },
   "dependencies": {
     "@date-io/date-fns": "^1.3.13",
+    "@fortawesome/fontawesome-svg-core": "^1.2.35",
+    "@fortawesome/free-brands-svg-icons": "^5.15.3",
+    "@fortawesome/free-regular-svg-icons": "^5.15.3",
+    "@fortawesome/free-solid-svg-icons": "^5.15.3",
+    "@fortawesome/react-fontawesome": "^0.1.14",
     "@material-ui/core": "^4.11.2",
     "@material-ui/icons": "^4.11.2",
+    "@material-ui/lab": "^4.0.0-alpha.57",
     "@material-ui/pickers": "^3.2.10",
     "bcrypt": "^5.0.1",
     "date-fns": "^2.19.0",
@@ -18,10 +24,15 @@
     "mongoose": "^5.11.16",
     "mui-datatables": "^3.7.6",
     "mui-rte": "^1.26.2",
-    "next": "10.0.3",
+    "next": "^10.1.3",
     "next-auth": "^3.11.2",
     "next-connect": "^0.10.0",
     "react": "17.0.1",
-    "react-dom": "17.0.1"
+    "react-dom": "17.0.1",
+    "swr": "^0.5.5"
+  },
+  "devDependencies": {
+    "eslint-config-google": "^0.14.0",
+    "eslint-plugin-react": "^7.23.2"
   }
 }
