@@ -1,10 +1,10 @@
-'use strict';
-const mongoose = require('mongoose');
+// "use strict";
+const mongoose = require("mongoose");
 
 const ProfileTypeModel = new mongoose.Schema({
-    name: {type: String, required: true, trim: true, unique: true},
-    status: {type: Boolean, required: true, default: true},
-    creationDate: {type: Date, default: Date.now},
+	name: { type: String, required: true, trim: true, unique: true },
+	status: { type: Boolean, required: true, default: true },
+	creationDate: { type: Date, default: Date.now },
 });
 
-module.exports = mongoose.model('ProfileType', ProfileTypeModel);
+module.exports = mongoose.models.ProfileType || mongoose.model("ProfileType", ProfileTypeModel);
