@@ -1,14 +1,14 @@
-import Head from 'next/head'
-import React from 'react';
-import Layout, {siteTittle} from '../../components/layout'
+import Head from "next/head";
+import React from "react";
+import { Layout, siteTittle } from "../../components";
 
-export default function Task(){
-  return (
-    <Layout>
-      <Head>
-        <title>{siteTittle}</title>
-      </Head>
-      <h1>This is Tasks</h1>
-    </Layout>
-  ) 
+export default function Task() {
+	return (
+		<Layout>
+			<Head>
+				<title>{siteTittle}</title>
+			</Head>
+			<h1>This is Tasks</h1>
+		</Layout>
+	);
 }
